backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai",
    "fastapi",
    "uvicorn",
    "python-docx",
    "google-generativeai",
    "httpx",
    "pydantic",
    "pydantic-settings",
    "python-multipart",
    "pytest",
    "pytest-asyncio",
    "langchain",
    "langgraph",
    "deepagents>=0.4.7",
    "langchain-openai",
    "python-dotenv",
    "sqlalchemy>=2.0.48",
    "aiosqlite>=0.22.1",
    "alembic>=1.18.4",
    "greenlet>=3.3.2",
]
